refactor rich text SCSS
