+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@contentful/rich-text-react-renderer": "^15.11.1",
+    "@contentful/rich-text-types": "^15.11.1",
     "next": "^12.0.9",
     "react": "17.0.2",
     "react-dom": "17.0.2",
